Split the too long kiwix-build.py file into several smaller ones.
